Синхронизирован код парсера с основной логикой приложения. Добавлен эндпоинт на получение всех документов из базы. Увеличен общий лимит памяти на одновременную загрузку документов.

+++ src/main/java/org/skeleton/team/service/UplpDocServiceImpl.java
@@ -1,17 +1,13 @@
 package org.skeleton.team.service;
 
 import lombok.RequiredArgsConstructor;
-import org.apache.pdfbox.cos.COSDocument;
-import org.apache.pdfbox.io.RandomAccessFile;
-import org.apache.pdfbox.pdfparser.PDFParser;
-import org.apache.pdfbox.pdmodel.PDDocument;
-import org.apache.pdfbox.text.PDFTextStripper;
 import org.skeleton.team.entity.UplpDoc;
 import org.skeleton.team.entity.UplpLog;
 import org.skeleton.team.entity.UplpSimpleDoc;
 import org.skeleton.team.mapper.UplpDocMapper;
 import org.skeleton.team.repository.UplpDocRepository;
 import org.skeleton.team.repository.UplpLogRepository;
+import org.skeleton.team.service.parser_dev.UPLPParser;
 import org.skeleton.team.util.UplpDocConverter;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -19,6 +15,7 @@ import org.springframework.web.multipart.MultipartFile;
 
 import javax.servlet.http.HttpServletRequest;
 import java.io.File;
+import java.io.IOException;
 import java.io.OutputStream;
 import java.sql.Timestamp;
 import java.util.*;
@@ -35,6 +32,7 @@ public class UplpDocServiceImpl implements UplpDocService {
     private final HttpServletRequest request;
     private final UplpDocConverter uplpDocConverter;
     private final UplpDocMapper uplpDocMapper;
+    private final UPLPParser uplpParser;
 
     /**
      * Получение документа ГПЗУ по ИД.
@@ -59,6 +57,16 @@ public class UplpDocServiceImpl implements UplpDocService {
     }
 
     /**
+     * Получение всех документов ГПЗУ.
+     * @return документы ГПЗУ
+     */
+    @Override
+    @Transactional(readOnly = true)
+    public List<UplpDoc> getAllUplpDocs() {
+        return uplpDocRepository.findAll();
+    }
+
+    /**
      * Получение простых документов ГПЗУ по ИД.
      * @param ids идентификаторы документов ГПЗУ
      * @return документы ГПЗУ в формате без системных полей
@@ -80,8 +88,9 @@ public class UplpDocServiceImpl implements UplpDocService {
         UplpDoc uplpDoc = uplpDocRepository.findById(id).orElse(null);
         if (uplpDoc != null) {
             String filePath = uplpDoc.getFileReference();
-            File file = new File(filePath);
-            file.delete();
+            if (!new File(filePath).delete()) {
+                throw new RuntimeException("Не удалось удалить файл из файловой системы");
+            }
             uplpDocRepository.delete(uplpDoc);
             return uplpDoc;
         }
@@ -98,44 +107,21 @@ public class UplpDocServiceImpl implements UplpDocService {
     public List<UplpDoc> createUplpDocs(List<MultipartFile> multipartFiles) {
         List<UplpDoc> result = new ArrayList<>();
         String realPathtoUploads = request.getServletContext().getRealPath("/files/");
-        if(! new File(realPathtoUploads).exists()) {
-            new File(realPathtoUploads).mkdir();
+        if(!new File(realPathtoUploads).exists()) {
+            if (!new File(realPathtoUploads).mkdir()) {
+                throw new RuntimeException("Не удалось создать директорию для хранения фалйов");
+            }
         }
 
-        for (int i = 0; i < multipartFiles.size(); i++) {
+        for (MultipartFile multipartFile : multipartFiles) {
             try {
-                String orgName = multipartFiles.get(i).getOriginalFilename();
+                String orgName = multipartFile.getOriginalFilename();
                 Date loadTime = new Timestamp(System.currentTimeMillis());
                 String filePath = realPathtoUploads + loadTime.getTime() + orgName;
                 File file = new File(filePath);
-                multipartFiles.get(i).transferTo(file);
-                String parsedText;
-                PDFParser parser = new PDFParser(new RandomAccessFile(file, "r"));
-                parser.parse();
-                COSDocument cosDoc = parser.getDocument();
-                PDFTextStripper pdfStripper = new PDFTextStripper();
-                PDDocument pdDoc = new PDDocument(cosDoc);
-                parsedText = pdfStripper.getText(pdDoc);
-                String[] words = parsedText.split(" ");
-                UplpDoc uplpDoc = new UplpDoc();
-                StringBuilder parseErrors = new StringBuilder();
-                for (String word : words) {
-                    //TODO парсим
-                }
-                UplpLog uplpLog = new UplpLog();
-                uplpLog.setFileName(orgName);
-                uplpLog.setLoadDttm(loadTime);
-                uplpLog.setProcessDttm(new Timestamp(System.currentTimeMillis()));
-                uplpDoc.setFileReference(file.getAbsolutePath());
-                if (parseErrors.length() == 0) {
-                    uplpLog.setProcessResult("Успешно");
-                } else {
-                    uplpLog.setProcessResult("Неуспешно");
-                    uplpLog.setErrors(parseErrors.toString());
-                }
-                uplpLog = uplpLogRepository.save(uplpLog);
-                uplpDoc.setUplpLog(uplpLog);
-                result.add(uplpDocRepository.save(uplpDoc));
+                multipartFile.transferTo(file);
+
+                result.add(uplpDocRepository.save(parseFile(orgName, loadTime, file)));
             } catch (Exception e) {
                 throw new RuntimeException("Критическая ошибка обработки файлов: " + e.getMessage());
             }
@@ -152,10 +138,11 @@ public class UplpDocServiceImpl implements UplpDocService {
     @Override
     @Transactional
     public UplpDoc updateUplpDoc(MultipartFile multipartFile, Long id) {
-        //TODO поправить после изменений в создании документа
         String realPathtoUploads = request.getServletContext().getRealPath("/files/");
-        if(! new File(realPathtoUploads).exists()) {
-            new File(realPathtoUploads).mkdir();
+        if(!new File(realPathtoUploads).exists()) {
+            if (!new File(realPathtoUploads).mkdir()) {
+                throw new RuntimeException("Не удалось создать директорию для хранения фалйов");
+            }
         }
         try {
             deleteUplpDocById(id);
@@ -164,33 +151,8 @@ public class UplpDocServiceImpl implements UplpDocService {
             String filePath = realPathtoUploads + loadTime.getTime() + orgName;
             File file = new File(filePath);
             multipartFile.transferTo(file);
-            String parsedText;
-            PDFParser parser = new PDFParser(new RandomAccessFile(file, "r"));
-            parser.parse();
-            COSDocument cosDoc = parser.getDocument();
-            PDFTextStripper pdfStripper = new PDFTextStripper();
-            PDDocument pdDoc = new PDDocument(cosDoc);
-            parsedText = pdfStripper.getText(pdDoc);
-            String[] words = parsedText.split(" ");
-            UplpDoc uplpDoc = new UplpDoc();
-            StringBuilder parseErrors = new StringBuilder();
-            for (String word : words) {
-                //TODO парсим
-            }
-            UplpLog uplpLog = new UplpLog();
-            uplpLog.setFileName(orgName);
-            uplpLog.setLoadDttm(loadTime);
-            uplpLog.setProcessDttm(new Timestamp(System.currentTimeMillis()));
-            uplpDoc.setFileReference(file.getAbsolutePath());
-            if (parseErrors.length() == 0) {
-                uplpLog.setProcessResult("Успешно");
-            } else {
-                uplpLog.setProcessResult("Неуспешно");
-                uplpLog.setErrors(parseErrors.toString());
-            }
-            uplpLog = uplpLogRepository.save(uplpLog);
-            uplpDoc.setUplpLog(uplpLog);
-            return uplpDocRepository.save(uplpDoc);
+
+            return uplpDocRepository.save(parseFile(orgName, loadTime, file));
         } catch (Exception e) {
             throw new RuntimeException("Критическая ошибка обработки файлов: " + e.getMessage());
         }
@@ -218,5 +180,26 @@ public class UplpDocServiceImpl implements UplpDocService {
         List<UplpDoc> docs = getUplpDocByIds(ids);
         List<UplpSimpleDoc> simpleDocs = uplpDocMapper.toSimpleDoc(docs);
         uplpDocConverter.convertUplpDocsToXmlStream(simpleDocs, outputStream);
+    }
+
+    private UplpDoc parseFile(String fileName, Date loadTime, File file) throws IOException {
+        StringBuilder parseErrors = new StringBuilder();
+
+        UplpDoc uplpDoc = uplpParser.parsingUPLPFile(file, parseErrors);
+
+        UplpLog uplpLog = new UplpLog();
+        uplpLog.setFileName(fileName);
+        uplpLog.setLoadDttm(loadTime);
+        uplpLog.setProcessDttm(new Timestamp(System.currentTimeMillis()));
+        uplpDoc.setFileReference(file.getAbsolutePath());
+        if (parseErrors.length() == 0) {
+            uplpLog.setProcessResult("Успешно");
+        } else {
+            uplpLog.setProcessResult("Неуспешно");
+            uplpLog.setErrors(parseErrors.toString());
+        }
+        uplpLog = uplpLogRepository.save(uplpLog);
+        uplpDoc.setUplpLog(uplpLog);
+        return uplpDoc;
     }
 }
